// app/src/lib/types.ts
/*
 Generated by typeshare 1.0.0
*/

export interface ParseResult {
	parseTree: string;
	markers: MarkerData[];
}

export interface MarkerData {
	relatedInformation?: RelatedInformation[];
	tags?: MarkerTag[];
	severity: MarkerSeverity;
	message: string;
	startLineNumber: number;
	startColumn: number;
	endLineNumber: number;
	endColumn: number;
}

export interface RelatedInformation {
	/** Is actually a `Uri` */
	resource: string;
	message: string;
	startLineNumber: number;
	startColumn: number;
	endLineNumber: number;
	endColumn: number;
}

export enum MarkerSeverity {
	Hint = "Hint",
	Info = "Info",
	Warning = "Warning",
	Error = "Error",
}

export enum MarkerTag {
	Unnecessary = "Unnecessary",
	Deprecated = "Deprecated",
}